Move individual driver declarations out to each module.
